[Thermo] update unit tests
